requirements.txt
bs4
pillow
PyExecJS
requests
aiowebsocket
plotly
flask
PyYAML